+++ src/stores/ActivityStore/Erc20Token.ts
@@ -1,7 +1,7 @@
 import { createERC20Contract, getTokenConfig } from '@cere-wallet/wallet-engine';
 import { BigNumber, utils } from 'ethers';
 
-import { Wallet } from '../types';
+import { ReadyWallet } from '../types';
 import { Activity, ActivityStore } from './ActivityStore';
 
 type Log = {
@@ -19,23 +19,38 @@ type Args = utils.Result & {
 
 export class Erc20Token {
   private tokenConfig = getTokenConfig();
-  private interface: utils.Interface;
+  private interface?: utils.Interface;
+  private dispose?: () => void;
 
-  constructor(private activityStore: ActivityStore, { provider, network, account }: Wallet) {
-    const erc20 = createERC20Contract(provider!.getSigner(), network!.chainId);
-    const receiveFilter = erc20.filters.Transfer(null, account!.address);
-    const sendFilter = erc20.filters.Transfer(account!.address);
+  constructor(private activityStore: ActivityStore) {}
 
-    provider!.on(receiveFilter, (log) => this.onReceive(log));
-    provider!.on(sendFilter, (log) => this.onSend(log));
+  start({ provider, network, account }: ReadyWallet) {
+    const erc20 = createERC20Contract(provider.getSigner(), network.chainId);
+    const receiveFilter = erc20.filters.Transfer(null, account.address);
+    const sendFilter = erc20.filters.Transfer(account.address);
+
+    provider.on(receiveFilter, this.onReceive);
+    provider.on(sendFilter, this.onSend);
 
     this.interface = erc20.interface;
+
+    this.dispose = () => {
+      provider.off(receiveFilter, this.onReceive);
+      provider.off(sendFilter, this.onSend);
+    };
+  }
+
+  stop() {
+    this.dispose?.();
+
+    this.interface = undefined;
+    this.dispose = undefined;
   }
 
   private createActivity(type: Activity['type'], log: Log): Activity {
     const { symbol, decimals } = this.tokenConfig;
     const { transactionHash } = log;
-    const { from, to, value } = this.interface.parseLog(log).args as Args;
+    const { from, to, value } = this.interface!.parseLog(log).args as Args;
     const amount = value.div(10 ** decimals).toNumber();
 
     return {
@@ -56,15 +71,15 @@ export class Erc20Token {
     };
   }
 
-  private onReceive(log: Log) {
+  private onReceive = (log: Log) => {
     const activity = this.createActivity('in', log);
 
     this.activityStore.addActivity(activity);
-  }
+  };
 
-  private onSend(log: Log) {
+  private onSend = (log: Log) => {
     const activity = this.createActivity('out', log);
 
     this.activityStore.addActivity(activity);
-  }
+  };
 }