+++ src/stores/NetworkStore/NetworkStore.ts
@@ -1,4 +1,4 @@
-import { ChainConfig } from '@cere-wallet/communication';
+import { ChainConfig, getChainConfig, DEFAULT_NETWORK } from '@cere-wallet/communication';
 import { makeAutoObservable } from 'mobx';
 
 import { Wallet } from '../types';
@@ -8,11 +8,15 @@ type SharedState = {
   network?: ChainConfig;
 };
 
+const defaultNetwork = getChainConfig(DEFAULT_NETWORK);
+
 export class NetworkStore {
   private shared = createSharedState<SharedState>(
     `network.${this.wallet.instanceId}`,
-    {},
-    { readOnly: !this.wallet.isRoot },
+    {
+      network: defaultNetwork,
+    },
+    { readOnly: !this.wallet.isRoot() },
   );
 
   constructor(private wallet: Wallet) {