+++ packages/communication/src/index.ts
@@ -3,3 +3,5 @@ export * from './rpc';
 export * from './popup';
 
 export * from './getIFrameOrigin';
+export * from './getChainConfig';
+export * from './constants';