+++ src/hooks/useAuthenticationStore.ts
@@ -1,0 +1,7 @@
+import { useWalletStore } from './useWalletStore';
+
+export const useAuthenticationStore = () => {
+  const walletStore = useWalletStore();
+
+  return walletStore.authenticationStore;
+};