+++ src/stores/AssetStore/Erc20Token.ts
@@ -2,19 +2,19 @@ import { makeAutoObservable } from 'mobx';
 import { fromResource } from 'mobx-utils';
 import { createERC20Contract, getTokenConfig, TokenConfig } from '@cere-wallet/wallet-engine';
 
-import { Asset, Wallet } from '../types';
+import { Asset, ReadyWallet } from '../types';
 
-const createBalanceResource = ({ provider, network, account }: Wallet, { decimals }: TokenConfig) => {
+const createBalanceResource = ({ provider, network, account }: ReadyWallet, { decimals }: TokenConfig) => {
   let currentListener: () => {};
 
-  const erc20 = createERC20Contract(provider!.getSigner(), network!.chainId);
-  const receiveFilter = erc20.filters.Transfer(null, account!.address);
-  const sendFilter = erc20.filters.Transfer(account!.address);
+  const erc20 = createERC20Contract(provider.getSigner(), network.chainId);
+  const receiveFilter = erc20.filters.Transfer(null, account.address);
+  const sendFilter = erc20.filters.Transfer(account.address);
 
   return fromResource<number>(
     (sink) => {
       currentListener = async () => {
-        const balance = await erc20.balanceOf(account!.address);
+        const balance = await erc20.balanceOf(account.address);
 
         sink(balance.div(10 ** decimals).toNumber());
       };
@@ -41,7 +41,7 @@ export class Erc20Token implements Asset {
   private tokenConfig = getTokenConfig();
   private balanceResource = createBalanceResource(this.wallet, this.tokenConfig);
 
-  constructor(private wallet: Wallet) {
+  constructor(private wallet: ReadyWallet) {
     makeAutoObservable(this);
   }
 
@@ -50,7 +50,7 @@ export class Erc20Token implements Asset {
   }
 
   get network() {
-    return this.wallet.network!.displayName;
+    return this.wallet.network.displayName;
   }
 
   get ticker() {