+++ src/stores/WalletStore/WalletStore.ts
@@ -1,45 +1,68 @@
 import { randomBytes } from 'crypto';
 import { providers } from 'ethers';
-import { makeAutoObservable, when } from 'mobx';
-import { createProvider } from '@cere-wallet/wallet-engine';
+import { makeAutoObservable, runInAction, when } from 'mobx';
+import { createProvider, Provider as EngineProvider } from '@cere-wallet/wallet-engine';
 
-import { Provider, Wallet } from '../types';
+import { Provider, Wallet, WalletStatus } from '../types';
 import { AccountStore } from '../AccountStore';
 import { NetworkStore } from '../NetworkStore';
 import { AssetStore } from '../AssetStore';
 import { BalanceStore } from '../BalanceStore';
 import { ActivityStore } from '../ActivityStore';
 import { AppContextStore } from '../AppContextStore';
+import { AuthenticationStore } from '../AuthenticationStore';
 
 export class WalletStore implements Wallet {
-  readonly isRoot = false;
-
+  readonly instanceId: string;
   readonly accountStore: AccountStore;
   readonly networkStore: NetworkStore;
   readonly assetStore: AssetStore;
   readonly balanceStore: BalanceStore;
   readonly activityStore: ActivityStore;
+  readonly authenticationStore: AuthenticationStore;
   readonly appContextStore: AppContextStore;
 
   private currentProvider?: Provider;
+  private initialized = false;
+  private isRootInstance = false;
 
-  constructor(readonly instanceId: string = randomBytes(16).toString('hex')) {
+  constructor(instanceId?: string) {
     makeAutoObservable(this);
+
+    this.isRootInstance = !instanceId;
+    this.instanceId = instanceId || randomBytes(16).toString('hex');
 
     this.networkStore = new NetworkStore(this);
     this.accountStore = new AccountStore(this);
     this.assetStore = new AssetStore(this);
-    this.balanceStore = new BalanceStore(this.assetStore);
+    this.balanceStore = new BalanceStore(this, this.assetStore);
     this.activityStore = new ActivityStore(this);
     this.appContextStore = new AppContextStore(this);
+    this.authenticationStore = new AuthenticationStore(this.accountStore, this.appContextStore);
+  }
+
+  isRoot() {
+    return this.isRootInstance;
+  }
+
+  isReady() {
+    return !!(this.provider && this.network && this.account);
+  }
+
+  get status(): WalletStatus {
+    if (this.isReady()) {
+      return 'ready';
+    }
+
+    if (this.initialized && !this.account) {
+      return 'unauthenticated';
+    }
+
+    return 'errored';
   }
 
   get provider() {
     return this.currentProvider;
-  }
-
-  private set provider(provider) {
-    this.currentProvider = provider;
   }
 
   get network() {
@@ -51,12 +74,26 @@ export class WalletStore implements Wallet {
   }
 
   async init() {
-    await when(() => !!this.accountStore.account && !!this.networkStore.network);
+    let provider: EngineProvider | undefined;
 
-    const { privateKey } = this.accountStore.account!;
-    const chainConfig = this.networkStore.network!;
-    const provider = await createProvider({ privateKey, chainConfig });
+    await when(() => !!this.network);
 
-    this.provider = new providers.Web3Provider(provider);
+    if (this.isRoot()) {
+      await this.authenticationStore.rehydrate();
+    } else {
+      await when(() => !!this.account);
+    }
+
+    if (this.account && this.network) {
+      provider = await createProvider({
+        privateKey: this.account.privateKey,
+        chainConfig: this.network,
+      });
+    }
+
+    runInAction(() => {
+      this.currentProvider = provider && new providers.Web3Provider(provider);
+      this.initialized = true;
+    });
   }
 }