+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNext
 
 - [DAV-3126] Add wallet loading state + standalone login/logout
+- [DAV-3155] Add overlay modals support + enable by default in SDK
 
 ### v1.5.0